Added "dataset" module for generating test data as SQL table
Test flow works like this:

ruby script -> json file -> python script -> sql database

This is so we can still use the legacy app compatibility with the JSON
file as a debug database read in. The goal is to streamline this into a
single process - generate a test database with a single script. Broken
into Ruby/Python right now because Python’s Faker port doesn’t appear
as developed or stable as Ruby’s.